refactor(components): изменить структуру проекта
все страницы собраны в отдельной папке
