+++ src/components/Header/Header.test.js
@@ -1,7 +1,7 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { Header } from "./Header";
+import React from 'react';
+import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
+import { Header } from "./Header";
 
 it('render without crashing', () => {
    const div = document.createElement('div');
