+++ src/components/pages/Profile/Profile.module.css
@@ -4,7 +4,7 @@
 }
 
 .profile {
-   background: url(../../main-bg.png) center 50% no-repeat;
+   background: url(../../../main-bg.png) center 50% no-repeat;
 }
 
 .profile {
